Adds AbilityData ScriptableObject and removes AbilityNameStore text asset
