+++ Danki2/Assets/AbilityData.asset
@@ -1,0 +1,58 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2ad90acd62399ef41b142664a017c8ee, type: 3}
+  m_Name: New Ability Data
+  m_EditorClassIdentifier: 
+  abilityIds:
+  - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+  abilityDisplayNameDictionary:
+    _keys:
+    - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+    _values:
+    - 
+  abilityTypeDictionary:
+    _keys:
+    - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+    _values: 00000000
+  abilityDamageDictionary:
+    _keys:
+    - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+    _values: 00000000
+  abilityEmpowermentsDictionary:
+    _keys:
+    - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+    _values:
+    - empowerments: 
+  abilityRarityDictionary:
+    _keys:
+    - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+    _values: 00000000
+  abilityCollisionSoundLevelDictionary:
+    _keys:
+    - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+    _values: 00000000
+  abilityVocalisationTypeDictionary:
+    _keys:
+    - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+    _values: 00000000
+  abilityFmodEventDictionary:
+    _keys:
+    - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+    _values:
+    - fmodStartEventRef: 
+      fmodSwingEventRef: 
+      fmodImpactEventRef: 
+  abilityIconDictionary:
+    _keys:
+    - value: f7b337eb-4410-4d0c-826f-e09dc98da182
+    _values:
+    - {fileID: 0}
