+++ Danki2/Assets/AbilityData.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 29ec540ec70ed9c469aa954705eb26d9
-folderAsset: yes
-DefaultImporter:
+guid: 8277b108ff04a584985980857bd819d4
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
