+++ Danki2/Assets/Prefabs/Meta/GameplayManagers/AbilityLookup2.prefab
@@ -43,144 +43,4 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7fc2102be59865c48b30800c2844e753, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  abilityIds:
-  - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-  - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-  - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-  - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-  - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-  - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-  - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-  abilityDisplayNameDictionary:
-    _keys:
-    - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-    - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-    - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-    - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-    - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-    - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-    - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-    _values:
-    - Slash
-    - Smash
-    - Thrust
-    - Hack
-    - Thunderous Smash
-    - Skewer
-    - Calamity
-  abilityTypeDictionary:
-    _keys:
-    - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-    - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-    - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-    - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-    - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-    - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-    - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-    _values: 00000000010000000200000000000000010000000200000001000000
-  abilityDamageDictionary:
-    _keys:
-    - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-    - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-    - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-    - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-    - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-    - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-    - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-    _values: 0500000008000000060000000500000008000000050000000a000000
-  abilityEmpowermentsDictionary:
-    _keys:
-    - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-    - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-    - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-    - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-    - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-    - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-    - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-    _values:
-    - empowerments: 
-    - empowerments: 
-    - empowerments: 
-    - empowerments: 01000000
-    - empowerments: 00000000
-    - empowerments: 02000000
-    - empowerments: 0100000002000000
-  abilityRarityDictionary:
-    _keys:
-    - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-    - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-    - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-    - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-    - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-    - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-    - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-    _values: 00000000000000000000000001000000010000000100000002000000
-  abilityCollisionSoundLevelDictionary:
-    _keys:
-    - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-    - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-    - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-    - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-    - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-    - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-    - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-    _values: 00000000010000000000000000000000010000000000000001000000
-  abilityVocalisationTypeDictionary:
-    _keys:
-    - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-    - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-    - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-    - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-    - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-    - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-    - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-    _values: 02000000020000000200000001000000010000000200000001000000
-  abilityFmodEventDictionary:
-    _keys:
-    - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-    - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-    - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-    - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-    - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-    - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-    - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-    _values:
-    - fmodStartEventRef: 
-      fmodSwingEventRef: event:/AS_AbilitySounds/AS_Slash
-      fmodImpactEventRef: 
-    - fmodStartEventRef: event:/AS_AbilitySounds/AS_Smash/AS_Smash_Draw
-      fmodSwingEventRef: 
-      fmodImpactEventRef: event:/AS_AbilitySounds/AS_Smash/AS_Smash_Impact
-    - fmodStartEventRef: 
-      fmodSwingEventRef: 
-      fmodImpactEventRef: 
-    - fmodStartEventRef: 
-      fmodSwingEventRef: 
-      fmodImpactEventRef: 
-    - fmodStartEventRef: 
-      fmodSwingEventRef: 
-      fmodImpactEventRef: 
-    - fmodStartEventRef: 
-      fmodSwingEventRef: 
-      fmodImpactEventRef: 
-    - fmodStartEventRef: 
-      fmodSwingEventRef: 
-      fmodImpactEventRef: 
-  abilityIconDictionary:
-    _keys:
-    - value: 3a819a43-228d-47c6-ba3c-515ee21824eb
-    - value: 463df743-c237-4f9e-9b86-e1dac9a4d94b
-    - value: 16378d15-5980-44aa-85cc-c224ff2174d2
-    - value: d4298ed9-9bf2-44c8-a69b-dc5453b8718a
-    - value: 08e8d43d-8066-4cf2-9a89-21fd424b80d3
-    - value: b5cc5d27-12fe-4519-a93e-d10604bef5eb
-    - value: 593b5246-0a14-4ddd-b0a9-3702d72e92d3
-    _values:
-    - {fileID: 21300000, guid: 545702d9040f5024e9eed6cf8ac19a16, type: 3}
-    - {fileID: 21300000, guid: 77ca6265971ef524b8f098606670c3cf, type: 3}
-    - {fileID: 21300000, guid: ac04ca413322cf74c9d945688586923f, type: 3}
-    - {fileID: 21300000, guid: b6e6aa96fdf88654983c156e1dabe170, type: 3}
-    - {fileID: 21300000, guid: f38b8a88882a44649907521872d37df8, type: 3}
-    - {fileID: 21300000, guid: c6827bb6e30fc0142bdf085b7818719a, type: 3}
-    - {fileID: 21300000, guid: 46dffb3410840a646800a6a02fbda4f3, type: 3}
-  abilityNameStore: {fileID: 4900000, guid: bb542debb62b3234fa4a5982c5ab98c4, type: 3}
+  abilityData: {fileID: 11400000, guid: 8277b108ff04a584985980857bd819d4, type: 2}
