+++ Danki2/Assets/Prefabs/Meta/MapGenerationLookup.prefab
@@ -46,7 +46,7 @@ MonoBehaviour:
   abilityChoices: 3
   runeSockets: 3
   leftStartingAbilityName: Slash
-  rightStartingAbilityName: Smash
+  rightStartingAbilityName: Slash
   generatedRoomDepth: 2
   minRoomExits: 1
   maxRoomExits: 3
@@ -113,4 +113,4 @@ MonoBehaviour:
     - spawnedEnemies: 0100000001000000
     - spawnedEnemies: 02000000
     - spawnedEnemies: 02000000
-  abilityNameStore: {fileID: 4900000, guid: bb542debb62b3234fa4a5982c5ab98c4, type: 3}
+  abilityData: {fileID: 11400000, guid: 8277b108ff04a584985980857bd819d4, type: 2}
