+++ Danki2/Assets/Scripts/Abilities2/Lookup/AbilityData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ad90acd62399ef41b142664a017c8ee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
