+++ Danki2/Assets/Scripts/Abilities2/Lookup/Editor/AbilityDataEditor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 6c245b32980c4aecac4048e5cce4ea5a
+timeCreated: 1633273593
