--- Danki2/Assets/Scripts/Abilities2/Lookup/Editor/AbilityLookup2Editor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: a1ece90451fc41f1a64539d01c8ea32c
-timeCreated: 1630164585
