+++ Danki2/Assets/Scripts/Persistence/DevPersistenceManager.cs
@@ -33,7 +33,7 @@ public class DevPersistenceManager : PersistenceManager
     [SerializeField] public RoomType roomType = RoomType.Combat;
     [SerializeField] public List<string> abilityChoiceNames = new List<string>();
     [SerializeField] public bool hasHealed = false;
-    [SerializeField] public TextAsset abilityNameStore = null;
+    [SerializeField] public AbilityData abilityData = null;
 
     public override SaveData SaveData => GenerateNewSaveData();
