--- Danki2/Assets/TextAssets/AbilityNameStore.txt
@@ -1,1 +1,0 @@
-Slash,Smash,Thrust,Hack,Thunderous Smash,Skewer,Calamity
