--- Danki2/Assets/TextAssets/AbilityNameStore.txt.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bb542debb62b3234fa4a5982c5ab98c4
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
